Updated API and implemented support for $inlinecount and $top
Get functions now return a dictionary with keys for results,
errors, and the count returned by inlinecount.

Added support for inlinecount, as well as convenience paramater
count=True for get_documents to add inlinecount to query string.

Updated tests to match new get return values, as well as
tests for inlinecount and top